Break Up Auth Table File
